# Implement fftconvolve in place of binary_dilate for MeasureObjectNeighbors