--- a/cellprofiler/modules/measureobjectneighbors.py
+++ b/cellprofiler/modules/measureobjectneighbors.py
@@ -501,7 +501,10 @@ previously discarded objects.""".format(
                 # Find the neighbors
                 #
                 patch_mask = patch == (index + 1)
-                extended = scipy.signal.fftconvolve(patch_mask, strel, mode='same') > 0.5
+                if distance <= 5:
+                    extended = scind.binary_dilation(patch_mask, strel)
+                else:
+                    extended = scipy.signal.fftconvolve(patch_mask, strel, mode='same') > 0.5
                 neighbors = np.unique(npatch[extended])
                 neighbors = neighbors[neighbors != 0]
                 if self.neighbors_are_objects:
@@ -536,9 +539,15 @@ previously discarded objects.""".format(
                     )
                 if self.neighbors_are_objects:
                     extendme = (patch != 0) & (patch != object_number)
-                    extended = scipy.signal.fftconvolve(extendme, strel_touching, mode='same') > 0.5
+                    if distance <= 5:
+                        extended = scind.binary_dilation(extendme, strel_touching)
+                    else:
+                        extended = scipy.signal.fftconvolve(extendme, strel_touching, mode='same') > 0.5
                 else:
-                    extended = scipy.signal.fftconvolve((npatch != 0), strel_touching, mode='same') > 0.5
+                    if distance <= 3:
+                        extended = scind.binary_dilation((npatch != 0), strel_touching)
+                    else:
+                        extended = scipy.signal.fftconvolve((npatch != 0), strel_touching, mode='same') > 0.5
                 overlap = np.sum(outline_patch & extended)
                 pixel_count[index] = overlap
             if sum([len(x) for x in first_objects]) > 0: