@DavidStirling 2020-06-26
@0x00b1 I've now made it so that the original method is used when the expansion distance is small (<=5). That seems to be a reasonable cutoff for where convolve method becomes faster.